Create useStartChatWithFriend and useRemoveFrined
